refact: move table container variables to its own file
